refactor: rename http users remote data source implementation
